backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
passlib
pyjwt
email-validator
pydantic[email]
